Update MOM6 to v1.1.0
This is a companion PR to https://github.com/GEOS-ESM/GEOSgcm_GridComp/pull/339 from @sanAkel for `components.yaml`. 

I'll defer to him as to whether this is 0-diff or not as https://github.com/GEOS-ESM/GEOSgcm_GridComp/pull/339 does not have a label either. Also, I'm not sure if updating to MOM6 1.1.0 will cause a failure without the changes in https://github.com/GEOS-ESM/GEOSgcm_GridComp/pull/339