Update to ESMA_cmake v3.44.0 and GEOS_Util v2.0.8
This PR updates the SLES15 detection such that if the
model is not built on SLES15, then the `BUILT_ON_SLES15` CMake
variable will now be `FALSE` instead of a blank string.

This PR also then updates to GEOS_Util v2.0.8 which has a fix
for `remap_restarts.py` to work with the new `BUILT_ON_SLES15` being
`FALSE`.